modified: return value dependency, and modified:change eventNumgerType int to long

+++ src/jp/ac/titech/cs/sa/tklab/faultlocalize/ppdebugger/creator/VariableReferenceVisitor.java
@@ -28,7 +28,7 @@ public class VariableReferenceVisitor {
 		
 		//データ依存を作成
 		DataDependency dd = new DataDependency(NameCreator.createVariableName(ref, scope),fromSd);
-		DataDependencySet dds = new DataDependencySet(currentSd,dd,Integer.valueOf(ref.getEventNumber()));
+		DataDependencySet dds = new DataDependencySet(currentSd,dd,Long.valueOf(ref.getEventNumber()));
 		
 		//依存先のステートメントにデータ依存を追加
 		em.addDataDependencySet(currentSd, dds);
